backend/requirements.txt
google-genai
python-dotenv
langchain
langchain-community
langchain-text-splitters
fastapi
uvicorn
qdrant-client
pypdf
sentence-transformers
pydantic
pymongo